openai-python-sdk/requirements.txt
openai>=1.57.0,<2.0.0
python-dotenv>=1.0.1,<2.0.0
agentops>=0.3.21,<0.4.0
packaging>=24.2